+++ xenian_channel/bot/models/__init__.py
@@ -9,3 +9,4 @@ from .tg_chat import *
 from .tg_user import *
 from .tg_message import *
 from .user_state import *
+from .button import *